// generatedTypes/accountDataSchema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type AccountDataSchema = [
  {
    bankNumber: {
      [k: string]: unknown;
    };
    extendedBankNumber: {
      [k: string]: unknown;
    };
    branchNumber: number;
    accountNumber: number;
    partyPreferredIndication: {
      [k: string]: unknown;
    };
    partyAccountInvolvementCode: 1 | 201;
    accountDealDate: {
      [k: string]: unknown;
    };
    accountUpdateDate: {
      [k: string]: unknown;
    };
    metegDoarNet: {
      [k: string]: unknown;
    };
    kodHarshaatPeilut: {
      [k: string]: unknown;
    };
    accountClosingReasonCode: {
      [k: string]: unknown;
    };
    /**
     * Branch number with account number and two spaces in between
     */
    productLabel: string;
    accountAgreementOpeningDate: {
      [k: string]: unknown;
    };
    serviceAuthorizationDesc: {
      [k: string]: unknown;
    };
    branchTypeCode: {
      [k: string]: unknown;
    };
    mymailEntitlementSwitch: {
      [k: string]: unknown;
    };
  }
];